[scheduler] Make all non-ghost workspaces replace ghosts
