+++ components/ee/ws-scheduler/pkg/scheduler/scheduler_test.go
@@ -291,7 +291,7 @@ func TestRequiredServices(t *testing.T) {
 
 			ctx, cancel := context.WithCancel(context.Background())
 			scheduler.startInformer(ctx)
-			state, err := scheduler.buildState(ctx, test.TargetPod, IsRegularWorkspace(test.TargetPod))
+			state, err := scheduler.buildState(ctx, test.TargetPod, IsNonGhostWorkspace(test.TargetPod))
 			cancel()
 			if err != nil {
 				t.Fatal(err)
